Give the two config-trim copies one home, and fix the dead one
tests/generate_intermediates.py and tests/conftest.py each carried their own
_trim_label_configs. The conftest copy globs config/training/fwd_*.yaml and
works; the generate_intermediates copy globbed config/fwd_*.yaml, where no
training menu has ever lived, so it matched nothing and every fixture
regeneration ran the full sweep. Measured on fx_pairs: linear 28 configs and
gbm 15 before, 2 and 2 after.

preset_patches.py already exists for exactly this - it holds the workload
reduction table both modules import, because two copies of it had previously
drifted. The trim belongs there for the same reason, so it moves there and
both call sites import it.

No CI behaviour changes: conftest keeps the function it already had, now by
import. Only fixture regeneration, which ran unreduced, is affected.
